pyproject.toml
[project]
name = "pdf2md"
version = "0.1.0"
description = "Convert PDF files to structured Markdown for LLM consumption"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "marker-pdf>=1.0.0",
    "click>=8.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "truststore>=0.9",
    "platformdirs>=3.0",
]

[project.optional-dependencies]
tiktoken = ["tiktoken>=0.5"]

[project.scripts]
pdf2md = "pdf2md.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf2md"]

[tool.uv]
index-url = "https://artifact.lhv.eu/artifactory/api/pypi/pypi/simple"

[dependency-groups]
dev = [
    "fpdf2>=2.7",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
markers = ["slow: requires ML models"]
